version 4

+++ izolabella.OpenWeatherMap.NET/Classes/Internals/Main.cs
@@ -9,7 +9,7 @@
         /// <summary>
         /// Initialize a new instance of the <see cref="Main"/> class.
         /// </summary>
-        public Main(decimal FeelsLike, decimal Temp, decimal TempMinimum, decimal TempMaximum, int Pressure, int Humidity)
+        public Main(double FeelsLike, double Temp, double TempMinimum, double TempMaximum, int Pressure, int Humidity)
         {
             this.FeelsLike = FeelsLike;
             this.Temp = Temp;
@@ -24,25 +24,25 @@
         /// The current temperature in Celsius as perceived by a human.
         /// </summary>
         [JsonProperty("feels_like")]
-        public decimal FeelsLike { get; set; }
+        public double FeelsLike { get; set; }
 
         /// <summary>
         /// The current temperature in Celsius.
         /// </summary>
         [JsonProperty("temp")]
-        public decimal Temp { get; set; }
+        public double Temp { get; set; }
 
         /// <summary>
         /// The minimum temperature currently observed in Celsius, unless this was returned from a forecast: the minimum temperature forecasted in Celsius.
         /// </summary>
         [JsonProperty("temp_min")]
-        public decimal TempMinimum { get; set; }
+        public double TempMinimum { get; set; }
 
         /// <summary>
         /// The maximum temperature currently observed in Celsius, unless this was returned from a forecast: the maximum temperature forecasted in Celsius.
         /// </summary>
         [JsonProperty("temp_max")]
-        public decimal TempMaximum { get; set; }
+        public double TempMaximum { get; set; }
 
         /// <summary>
         /// Atmospheric pressure (on the sea level, if there is no sea_level or grnd_level data), hPa
